# [extend #1191] Unnecessary alloc for XML, JSON, JSONP