--- a/context.go
+++ b/context.go
@@ -423,9 +423,9 @@ func (c *context) jsonPBlob(code int, callback string, i interface{}) (err error
 	return
 }
 
-func (c *context) jsonBlob(code int, i interface{}, pretty bool, indent string) error {
+func (c *context) jsonBlob(code int, i interface{}, indent string) error {
 	enc := json.NewEncoder(c.response)
-	if pretty {
+	if indent != "" {
 		enc.SetIndent("", indent)
 	}
 	c.writeContentType(MIMEApplicationJSONCharsetUTF8)
@@ -439,14 +439,13 @@ func (c *context) JSON(code int, i interface{}) (err error) {
 		indent string
 	)
 	if _, pretty = c.QueryParams()["pretty"]; c.echo.Debug || pretty {
-		pretty = true
 		indent = "  "
 	}
-	return c.jsonBlob(code, i, pretty, indent)
+	return c.jsonBlob(code, i, indent)
 }
 
 func (c *context) JSONPretty(code int, i interface{}, indent string) (err error) {
-	return c.jsonBlob(code, i, true, indent)
+	return c.jsonBlob(code, i, indent)
 }
 
 func (c *context) JSONBlob(code int, b []byte) (err error) {
@@ -470,11 +469,11 @@ func (c *context) JSONPBlob(code int, callback string, b []byte) (err error) {
 	return
 }
 
-func (c *context) xmlBlob(code int, i interface{}, pretty bool, indent string) (err error) {
+func (c *context) xmlBlob(code int, i interface{}, indent string) (err error) {
 	c.writeContentType(MIMEApplicationXMLCharsetUTF8)
 	c.response.WriteHeader(code)
 	enc := xml.NewEncoder(c.response)
-	if pretty {
+	if indent != "" {
 		enc.Indent("", "  ")
 	}
 	if _, err = c.response.Write([]byte(xml.Header)); err != nil {
@@ -489,14 +488,13 @@ func (c *context) XML(code int, i interface{}) (err error) {
 		indent string
 	)
 	if _, pretty = c.QueryParams()["pretty"]; c.echo.Debug || pretty {
-		pretty = true
 		indent = "  "
 	}
-	return c.xmlBlob(code, i, pretty, indent)
+	return c.xmlBlob(code, i, indent)
 }
 
 func (c *context) XMLPretty(code int, i interface{}, indent string) (err error) {
-	return c.xmlBlob(code, i, true, indent)
+	return c.xmlBlob(code, i, indent)
 }
 
 func (c *context) XMLBlob(code int, b []byte) (err error) {