Implement methods for
- JSON
- JSONP
- XML

fixes for tests
cc @vishr @ribice

#### Benchmark:
```go
var testUser = user{1, "Jon Snow"}

func BenchmarkAllocJSONP(b *testing.B) {
	e := New()
	req := httptest.NewRequest(POST, "/", strings.NewReader(userJSON))
	rec := httptest.NewRecorder()
	c := e.NewContext(req, rec).(*context)

	b.ResetTimer()
	b.ReportAllocs()

	for i := 0; i < b.N; i++ {
		c.JSONP(http.StatusOK, "callback", testUser)
	}
}

func BenchmarkAllocJSON(b *testing.B) {
	e := New()
	req := httptest.NewRequest(POST, "/", strings.NewReader(userJSON))
	rec := httptest.NewRecorder()
	c := e.NewContext(req, rec).(*context)

	b.ResetTimer()
	b.ReportAllocs()

	for i := 0; i < b.N; i++ {
		c.JSON(http.StatusOK, testUser)
	}
}

func BenchmarkAllocXML(b *testing.B) {
	e := New()
	req := httptest.NewRequest(POST, "/", strings.NewReader(userJSON))
	rec := httptest.NewRecorder()
	c := e.NewContext(req, rec).(*context)

	b.ResetTimer()
	b.ReportAllocs()

	for i := 0; i < b.N; i++ {
		c.XML(http.StatusOK, testUser)
	}
}
```

#### Result
```
# git checkout PR-branch
→ go test -test.v -test.run ^$ -test.bench ^BenchmarkAlloc* -count 10 > new.txt
# git checkout master-branch
→ go test -test.v -test.run ^$ -test.bench ^BenchmarkAlloc* -count 10 > old.txt

→ benchstat old.txt new.txt
name          old time/op    new time/op    delta
AllocJSONP-8    1.98µs ±17%    1.50µs ±12%  -24.11%  (p=0.000 n=10+9)
AllocJSON-8     1.82µs ±20%    1.73µs ± 6%     ~     (p=0.481 n=10+10)
AllocXML-8      4.27µs ±27%    4.27µs ±10%     ~     (p=0.549 n=10+9)

name          old alloc/op   new alloc/op   delta
AllocJSONP-8      193B ± 0%      162B ± 0%  -16.06%  (p=0.000 n=10+10)
AllocJSON-8       174B ± 0%      143B ± 0%  -17.82%  (p=0.000 n=10+10)
AllocXML-8      4.88kB ± 0%    4.77kB ± 0%   -2.30%  (p=0.019 n=6+8)

name          old allocs/op  new allocs/op  delta
AllocJSONP-8      5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
AllocJSON-8       3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=10+10)
AllocXML-8        11.0 ± 0%      10.0 ± 0%   -9.09%  (p=0.000 n=10+10)
```