@ribice 2018-10-11
Looks good to me. Thanks for the benchmarks!