@im-kulikov 2018-12-18
I just reopen, as I can see, all tests passed and has no conflicts with master..